requirements.txt
ipykernel
pydantic
langchain
langchain-community
python-dotenv
langchain-groq
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
mistralai
chromadb
faiss-cpu
langchain-chroma 
streamlit
fastapi
uvicorn
starlette
langserve
sse_starlette 